netz: fix crash when writing restart files.
the issue was that we were only nulling out the file pointer if an
error occurred when we closed it.  we'd then check if it was null
or not to decide opening a file, when the next open came, and it
wasn't null so we don't reopen.. but it wasn't valid.  eventually
this leads to badness.