netz: change config file format for slices.
now we expect a

  slices = [ z=5 x=6 y=9 ]

etc. line as the first line of psiphi.cfg.  This details the
slices to write out for *all* variables.  Users must give an empty
list if they do not desire outputs.

Furthermore, the 'slice=N' syntax is no longer valid within an
individual variable's specification.

This is still buggy in that it does not actually change which
dimension you set.  That is, it treats any 'A=N' specification as
if the user typed 'z=N'.