working solution for the '.mpi.c' problem.
Thanks to Nick Bowler on the automake list.  We need to use
".mpic" instead of ".mpi.c" for the extension, now, because suffix
rules in make suck.  But otherwise, works the same as before:
.mpic get compiled via the mpi compiler, whereas regular .c files
go through the normal C compiler.

One caveat is that we had to hack the c99 option into our mpicc
compilation rule, which hardcodes into GNU-land again.  Will need
to figure out an autoconf workaround for that.